Moved DatasetIndexer to modules
