+++ backend/dataall/api/Objects/Glossary/registry.py
@@ -6,7 +6,7 @@ from opensearchpy import OpenSearch
 from dataall.api import gql
 from dataall.api.gql.graphql_union_type import UnionTypeRegistry
 from dataall.db import Resource, models
-from dataall.searchproxy.indexers import DashboardIndexer, DatasetTableIndexer, DatasetIndexer
+from dataall.searchproxy.indexers import DashboardIndexer, DatasetTableIndexer
 from dataall.searchproxy.upsert import BaseIndexer
 
 
@@ -74,10 +74,4 @@ GlossaryRegistry.register(GlossaryDefinition(
     object_type="DatasetTable",
     model=models.DatasetTable,
     reindexer=DatasetTableIndexer
-))
-GlossaryRegistry.register(GlossaryDefinition(
-    target_type="Dataset",
-    object_type="Dataset",
-    model=models.Dataset,
-    reindexer=DatasetIndexer
 ))
