+++ tests/api/conftest.py
@@ -30,7 +30,7 @@ def patch_es(module_mocker):
     module_mocker.patch('dataall.searchproxy.search', return_value={})
     module_mocker.patch('dataall.searchproxy.upsert', return_value={})
     module_mocker.patch('dataall.searchproxy.indexers.upsert_dataset_tables', return_value={})
-    module_mocker.patch('dataall.searchproxy.indexers.DatasetIndexer.upsert', return_value={})
+    module_mocker.patch('dataall.modules.datasets.indexers.dataset_indexer.DatasetIndexer.upsert', return_value={})
     module_mocker.patch('dataall.searchproxy.indexers.DatasetTableIndexer.upsert', return_value={})
     module_mocker.patch(
         'dataall.modules.datasets.indexers.location_indexer.DatasetLocationIndexer.upsert',
